Config: Extract ConfigInfo into own header
