+++ Source/Core/Common/CMakeLists.txt
@@ -4,6 +4,7 @@ set(SRCS
   ColorUtil.cpp
   CommonFuncs.cpp
   Config/Config.cpp
+  Config/ConfigInfo.cpp
   Config/Layer.cpp
   Config/Section.cpp
   Crypto/AES.cpp
@@ -29,8 +30,8 @@ set(SRCS
   PcapFile.cpp
   PerformanceCounter.cpp
   Profiler.cpp
-  SettingsHandler.cpp
   SDCardUtil.cpp
+  SettingsHandler.cpp
   StringUtil.cpp
   SymbolDB.cpp
   SysConf.cpp
